move authentication.js into new controllers folder
